Split the region search into its own class
